Created the UserHandler class and used it in index.ts
